fix: High idle TPS hit of capacitor banks
We were syncing the positions of graph members over the network which was expensive to compute whether it has changed or not.
Because the graph was already local, I've shaved off some of the unnecessary graph logic on the client and use this locally instead.

This may not be the permanent solution, but is good enough to remove the immediate problem.